Organize app into components and pages with constant header
