+++ .gitignore
@@ -6,4 +6,4 @@ node_modules
 .output
 .env
 dist
-assets/resume.private.json
+*/resume.private.json
