--- components/Content/index.vue
@@ -1,26 +1,0 @@
-<template>
-  <div class="grid grid-cols-11 justify-center w-full h-full items-center mb-5 letter:mb-0">
-    <BasicsSummary />
-
-    <section class="px-4 col-span-11 medium:col-span-6 order-2 medium:order-1 h-full print:col-span-6 print:order-1">
-      <UiSectionHeading :title="'projects'" />
-      <Projects />
-    </section>
-
-    <section class="px-4 col-span-11 medium:col-span-5 order-1 medium:order-2 h-full print:col-span-5 print:order-2">
-      <UiSectionHeading :title="'skills'" />
-      <Skills />
-    </section>
-
-    <section class="px-4 col-span-11 medium:col-span-6 order-3 h-full print:col-span-6">
-      <UiSectionHeading :title="'experience'" />
-      <Work />
-    </section>
-
-    <section class="px-4 col-span-11 medium:col-span-5 order-4 h-full print:col-span-5">
-      <UiSectionHeading :title="'education'" />
-      <Education />
-    </section>
-
-  </div>
-</template>
